+++ cosmospy/__init__.py
@@ -3,5 +3,6 @@ from cosmospy._wallet import generate_wallet as generate_wallet  # noqa: F401
 from cosmospy._wallet import privkey_to_address as privkey_to_address  # noqa: F401
 from cosmospy._wallet import privkey_to_pubkey as privkey_to_pubkey  # noqa: F401
 from cosmospy._wallet import pubkey_to_address as pubkey_to_address  # noqa: F401
+from cosmospy._wallet import seed_to_privkey as seed_to_privkey  # noqa: F401
 
 __version__ = "4.0.0"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT