+++ cosmospy/typing.py
@@ -12,6 +12,8 @@ SyncMode = Literal["sync", "async", "block"]
 
 
 class Wallet(TypedDict):
+    seed: str
+    derivation_path: str
     private_key: str
     public_key: str
     address: str